// src/models/staff-resp.ts
// tslint:disable
/**
 * geo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 员工响应
 * @export
 * @interface StaffResp
 */
export interface StaffResp {
    /**
     * 员工ID
     * @type {number}
     * @memberof StaffResp
     */
    id?: number;
    /**
     * 用户名
     * @type {string}
     * @memberof StaffResp
     */
    username?: string;
    /**
     * 微信号
     * @type {string}
     * @memberof StaffResp
     */
    wechat?: string;
    /**
     * 手机号
     * @type {string}
     * @memberof StaffResp
     */
    phone?: string;
    /**
     * 姓名
     * @type {string}
     * @memberof StaffResp
     */
    name?: string;
    /**
     * 角色
     * @type {string}
     * @memberof StaffResp
     */
    role?: StaffRespRoleEnum;
    /**
     * 门店ID
     * @type {number}
     * @memberof StaffResp
     */
    storeId?: number;
    /**
     * 队长ID
     * @type {number}
     * @memberof StaffResp
     */
    leaderId?: number;
    /**
     * 状态
     * @type {number}
     * @memberof StaffResp
     */
    status?: number;
    /**
     * 创建时间
     * @type {string}
     * @memberof StaffResp
     */
    createdAt?: string;
    /**
     * 更新时间
     * @type {string}
     * @memberof StaffResp
     */
    updatedAt?: string;
}

/**
    * @export
    * @enum {string}
    */
export enum StaffRespRoleEnum {
    ADMIN = 'ADMIN',
    TEAMLEADER = 'TEAM_LEADER',
    SALESMAN = 'SALESMAN'
}